I'm designing a bike shed.
I'm pretty sure it should be yellow, for visiblity and safety of riders.
But then again, orange acomplishes this goal too but doesn't clash with
our interior design. I'm sure the safety implications are negligable to
begin with, and after all if safety was the main goal we should
re-design the structure itself around visiblity. That's a great point
Jim, can we re-schedule a meeting for next tesday after we've all had
time to re-evaluate the structure itself? Yes, of course. The safety of
the code... errr... bike shed... is very important to us here.